chore: renames add_workout* to create_workout*
